made flask run work and added search
