--- templates/index.html
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Flask MySQL App</title>
-</head>
-<body>
-    <h1>Users</h1>
-    <ul>
-        {% for user in users %}
-        <li>{{ user.uid }} {{ user.first_name }} {{ user.last_name }}</li>
-        {% endfor %}
-    </ul>
-
-    <h2>Add User</h2>
-    <form action="/add" method="post">
-        <input type="text" name="uid" placeholder="waterloo id" required>
-        <input type="text" name="fname" placeholder="first name" required>
-        <input type="text" name="lname" placeholder="last name" required>
-        <input type="text" name="startyear" placeholder="start year" required>
-        <button type="submit">Add User</button>
-    </form>
-
-    <h1>Course Search</h1>
-    <form method="POST" action="{{ url_for('search') }}">
-        <input type="text" name="query" placeholder="Enter a course title">
-        <button type="submit">Search</button>
-    </form>
-
-    {% if message %}
-        <p>{{ message }}</p>
-    {% elif courses %}
-        <h2>Search Results for "{{ query }}"</h2>
-        <ul>
-        {% for course in courses %}
-            <li>
-                <h3>{{ course.course_name }}</h3>
-                <p>{{ course.course_code }}</p>
-                <form method="POST" action="{{ url_for('add_course') }}">
-                    <input type="hidden" name="course_code" value="{{ course.course_code }}">
-                    <button type="submit">Add Course</button>
-                </form>
-            </li>
-        {% endfor %}
-        </ul>
-    {% endif %}
-
-    {% if results %}
-        <h2>Search Results</h2>
-        <ul>
-        {% for item in results %}
-            <li>{{ item }}</li>
-        {% endfor %}
-        </ul>
-    {% endif %}
-
-</body>
-</html>
